requirements.txt
flask
requests
pystray
pillow
